+++ packages/browser-wallet-api-helpers/src/wallet-api-types.ts
@@ -3,6 +3,7 @@ import type {
     AccountTransactionSignature,
     AccountTransactionType,
     JsonRpcClient,
+    SchemaVersion,
 } from '@concordium/web-sdk';
 
 /**
@@ -10,6 +11,7 @@ import type {
  */
 export enum EventType {
     AccountChanged = 'accountChanged',
+    AccountDisconnected = 'accountDisconnected',
     ChainChanged = 'chainChanged',
 }
 
@@ -25,7 +27,8 @@ interface Listeners<T extends EventType, Args extends any[]> {
 }
 
 type EventListeners = Listeners<EventType.AccountChanged, [accountAddress: string]> &
-    Listeners<EventType.ChainChanged, [chain: string]>;
+    Listeners<EventType.ChainChanged, [chain: string]> &
+    Listeners<EventType.AccountDisconnected, [accountAddress: string]>;
 
 interface MainWalletApi {
     /**
@@ -36,6 +39,7 @@ interface MainWalletApi {
      * @param payload the payload of the transaction to be signed and sent.
      * @param parameters parameters for the initContract and updateContract transactions in JSON-like format.
      * @param schema schema used for the initContract and updateContract transactions to serialize the parameters. Should be base64 encoded.
+     * @param schemaVersion version of the schema provided. Must be supplied for schemas that use version 0 or 1, as they don't have the version embedded.
      */
     sendTransaction(
         accountAddress: string,
@@ -44,7 +48,8 @@ interface MainWalletApi {
             | AccountTransactionType.InitializeSmartContractInstance,
         payload: AccountTransactionPayload,
         parameters: Record<string, unknown>,
-        schema: string
+        schema: string,
+        schemaVersion?: SchemaVersion
     ): Promise<string>;
     /**
      * Sends a transaction to the Concordium Wallet and awaits the users action. Note that a header is not sent, and will be constructed by the wallet itself.