+++ packages/browser-wallet-api/src/wallet-api.ts
@@ -4,6 +4,7 @@ import {
     AccountTransactionSignature,
     AccountTransactionType,
     JsonRpcClient,
+    SchemaVersion,
 } from '@concordium/web-sdk';
 import { WalletApi as IWalletApi, EventType } from '@concordium/browser-wallet-api-helpers';
 import EventEmitter from 'events';
@@ -72,7 +73,8 @@ class WalletApi extends EventEmitter implements IWalletApi {
         type: AccountTransactionType,
         payload: AccountTransactionPayload,
         parameters?: Record<string, unknown>,
-        schema?: string
+        schema?: string,
+        schemaVersion?: SchemaVersion
     ): Promise<string> {
         const response = await this.sendMessage<string | undefined>(MessageType.SendTransaction, {
             accountAddress,
@@ -80,6 +82,7 @@ class WalletApi extends EventEmitter implements IWalletApi {
             payload: stringify(payload),
             parameters,
             schema,
+            schemaVersion,
         });
 
         if (!response) {