+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -9,6 +9,11 @@
 ### (Breaking) Changed
 
 -   Updated API of sendTransaction and signMessage to require the account address.
+-   Updated API to include an 'accountDisconnected' event.
+
+## 0.1.1
+
+-   sendTransaction can now take a 5th argument, which is the schema's version. This will allow V1 contract parameters to be serialized.
 
 ## 0.1.0